added fix properties

+++ src/Features/Advertisement/Education/education.table.js
@@ -1,5 +1,5 @@
 
-import pool from "../../Mysql/mysql.database.js"
+import pool from "../../../Mysql/mysql.database.js";
 
 export const createEducationTable = async function () {
     try {
@@ -7,7 +7,7 @@ export const createEducationTable = async function () {
 
         // Define your CREATE TABLE query
         const createTableQuery = `CREATE TABLE IF NOT EXISTS education (
-            id INT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
+            id BIGINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
             plan_id BIGINT UNSIGNED NOT NULL,        
             user_id BIGINT UNSIGNED NOT NULL,
